Partial Collision Detection and Jump Stubs
Whew. Ok, it took a bit of work, but I managed to implement collision
detection for fences against other fences. This is well-documented in
the code itself as well. For my own sanity, I added a few System Outs to
check my work and confirm that it was working. And now that it is
working, our current problem is that Mhos and the Smiley are still
generated on top of fences. The good news is that the code I have
written is reproducible in this regard, meaning that I can
reverse-engineer this code to provide collision detection for the Mhos
and the Smiley. When I started working today I realized there was no
jump command and I started trying to implement that, but I realized that
the lack of collision detection prevented me from doing so properly. As
such, there are a few stubs for jump commands here and there. I had no
problems running the code at this point.