We are DONE!!!
Ok, not completely, but the program is functioning 99.999% perfectly.
Mho AI is finished, with the caveat that in certain configurations some
mhos will sit still for a turn or two. The reason for this is that the
mhos are not told to move in order of how close or far they are from the
player, but honestly, the game works just as well like this and honestly
I don't think it's necessary (plus I'm tired). Plenty of bugs were
squashed along the way, solving the mystery of why mhos would sometimes
still overlap fences (killMhos needed to be called more than once), and
a SIT direction was added (because sit is not the same as none, when I
made that mistake all the mhos converged on me instantly for an instant
game over :P). At this point, we have more or less everything required
of us besides telling the player that it is their turn, and we can write
in our documentation that we made this choice because we felt it would
be better to let the player dictate the pace of the game. This commit
also does not include the prompt boxes, expect one in about 10 minutes.