Added the zip archive, updated the jar
I updated the jar with the most recent changes. After I did this, I
created a zip archive with our jar and java files inside it. This zip
archive should be one of the two files you upload to Schoology, the
other being the Google Doc which is still in progress.